service refactor, version implementation process
Moved the database interactions to the repository. Prepared logic in services to work with versions. Created separate repositories with sql code instead of orm

+++ server/nest/src/users/users.service.ts
@@ -7,38 +7,68 @@ import { UsersRepository } from './users.repository';
 @Injectable()
 export class UsersService {
   constructor(private readonly usersRepository: UsersRepository) {}
-  async register(data: RegisterUserDto) {
-    const user = await this.usersRepository.findByUsernameOrEmail(
-      data.username,
-      data.password,
-      false,
-    );
-    if (user)
-      throw new HttpException(
-        'User with that email or username already exist',
-        HttpStatus.CONFLICT,
+
+  async register(data: RegisterUserDto, version: string = 'v1.1') {
+    if (version === 'v1.1') {
+      const user = await this.usersRepository.findByUsernameOrEmail(
+        data.username,
+        data.password,
+        false,
       );
-    const hashedPassword = await hash(data.password, 15);
-    const newUser = await this.usersRepository.create({
-      ...data,
-      password: hashedPassword,
-    });
-    return { username: newUser.username, id: newUser.id };
+      if (user)
+        throw new HttpException(
+          'User with that email or username already exists',
+          HttpStatus.CONFLICT,
+        );
+
+      const hashedPassword = await hash(data.password, 15);
+      const newUser = await this.usersRepository.create({
+        ...data,
+        password: hashedPassword,
+      });
+
+      return { username: newUser.username, id: newUser.id };
+    } else if (version === 'v1.2') {
+    } else {
+      throw new HttpException(
+        'Invalid version specified',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
   }
 
-  async login(data: LoginUserDto) {
-    const user = await this.usersRepository.findByUsernameOrEmail(
-      data.username,
-      '',
-    );
-    if (!(await compare(data.password, user.password)))
-      throw new HttpException(
-        'Incorrect username and password combination',
-        HttpStatus.NOT_FOUND,
+  async login(data: LoginUserDto, version: string = 'v1.1') {
+    if (version === 'v1.1') {
+      const user = await this.usersRepository.findByUsernameOrEmail(
+        data.username,
+        '',
       );
-    return { username: data.username, id: user.id };
+
+      if (!(await compare(data.password, user.password)))
+        throw new HttpException(
+          'Incorrect username and password combination',
+          HttpStatus.NOT_FOUND,
+        );
+
+      return { username: data.username, id: user.id };
+    } else if (version === 'v1.2') {
+    } else {
+      throw new HttpException(
+        'Invalid version specified',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
   }
-  async delete(id: number) {
-    return await this.usersRepository.deleteById(id);
+
+  async delete(id: number, version: string = 'v1.1') {
+    if (version === 'v1.1') {
+      return await this.usersRepository.deleteById(id);
+    } else if (version === 'v1.2') {
+    } else {
+      throw new HttpException(
+        'Invalid version specified',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
   }
 }
